Refactor: now everything runs just fine
