+++ solid_app/helpers/factories.py
@@ -2,8 +2,8 @@ from abc import ABC, abstractmethod
 from typing import Dict, Any
 from uuid import uuid4
 from sqlmodel import Session
-from .db_sqlite.database import get_session
-from .db_sqlite.models import User, UserType, Account, AccountType, AccountStatus
+from database.database import get_session
+from database.models import User, UserType, Account, AccountType, AccountStatus
 from fastapi import Depends
 from decimal import Decimal
 
@@ -18,12 +18,14 @@ class UserFactory(ABC):
     ) -> User:
         pass
 
+
 class AccountFactory(ABC):
     @abstractmethod
     def create_account(
         self, account_data: Dict[str, Any], session: Session = Depends(get_session)
     ) -> Account:
         pass
+
 
 class ClientFactory(UserFactory):
     """
@@ -56,40 +58,42 @@ class ManagerFactory(UserFactory):
         session.refresh(user)
         return user
 
+
 class SavingsAccountFactory(AccountFactory):
     """
     Factory Method applied to create savings accounts on the Database.
     """
+
     def create_account(
         self, account_data: Dict[str, Any], session: Session = Depends(get_session)
     ) -> Account:
-        
         account = Account(
             account_id=uuid4(),
             document_id=account_data["document_id"],
             account_type=AccountType.SAVINGS,
             balance=Decimal("0"),
-            status=AccountStatus.ACTIVE
+            status=AccountStatus.ACTIVE,
         )
         session.add(account)
         session.commit()
         session.refresh(account)
         return account
-    
+
+
 class CheckingAccountFactory(AccountFactory):
     """
     Factory Method applied to create checking accounts on the Database.
     """
+
     def create_account(
         self, account_data: Dict[str, Any], session: Session = Depends(get_session)
     ) -> Account:
-       
         account = Account(
             account_id=uuid4(),
             document_id=account_data["document_id"],
             account_type=AccountType.CHECKING,
             balance=Decimal("0"),
-            status=AccountStatus.ACTIVE
+            status=AccountStatus.ACTIVE,
         )
         session.add(account)
         session.commit()
