+++ solid_app/pyproject.toml
@@ -1,0 +1,27 @@
+[tool.poetry]
+name = "solid-app"
+version = "0.1.0"
+description = ""
+authors = ["Your Name <you@example.com>"]
+readme = "README.md"
+packages = [{include = "solid_app"}]
+
+[tool.poetry.dependencies]
+python = "^3.10"
+sqlalchemy = "^2.0.40"
+fastapi = {extras = ["standard"], version = "^0.115.12"}
+sqlmodel = "^0.0.24"
+pytest = "^8.3.5"
+coverage = "^7.8.0"
+pre-commit = "^4.2.0"
+
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
+
+[tool.pytest.ini_options]
+pythonpath = [
+  "."
+]
+testpaths = ["test"]
