+++ solid_app/test/test_factories.py
@@ -1,29 +1,29 @@
 from uuid import UUID
-from solid_app.src.factories import ClientFactory, ManagerFactory
-from solid_app.src.db_sqlite.models import User, UserType
-from decimal import Decimal
+from helpers.factories import ClientFactory, ManagerFactory
+from database.models import User, UserType
 
 
 def test_client_factory_create_user(client_user, mock_session):
     """Test that ClientFactory creates a user with CLIENT type and is_staff=False."""
     factory = ClientFactory()
-    
+
     def side_effect(user):
         return user
-    
+
     mock_session.refresh.side_effect = side_effect
-    
+
     user = factory.create_user(client_user, mock_session)
-    
+
     mock_session.add.assert_called_once()
     mock_session.commit.assert_called_once()
     mock_session.refresh.assert_called_once()
-    
+
     assert user.user_type == UserType.CLIENT
     assert user.is_staff is False
     assert user.document_id == "12345678901"
     assert user.username == "Lucky Luke"
     assert user.email == "lucky_mail@example.com"
+
 
 def test_manager_factory_create_user(mock_session):
     """Test that ManagerFactory creates a user with MANAGER type and is_staff=True."""
@@ -32,25 +32,26 @@ def test_manager_factory_create_user(mock_session):
         "document_id": "12345678901",
         "username": "testmanager",
         "email": "manager@example.com",
-        "name": "Test Manager"
+        "name": "Test Manager",
     }
-    
+
     def side_effect(user):
         return user
-    
+
     mock_session.refresh.side_effect = side_effect
-    
+
     user = factory.create_user(user_data, mock_session)
-    
+
     mock_session.add.assert_called_once()
     mock_session.commit.assert_called_once()
     mock_session.refresh.assert_called_once()
-    
+
     assert user.user_type == UserType.MANAGER
     assert user.is_staff is True
     assert user.document_id == "12345678901"
     assert user.username == "testmanager"
     assert user.email == "manager@example.com"
+
 
 def test_client_factory_integration(db_session):
     """Integration test that ClientFactory actually saves a user to the database."""
@@ -60,18 +61,18 @@ def test_client_factory_integration(db_session):
         "document_id": "12345678901",
         "username": "testclient",
         "email": "test@example.com",
-        "name": "Test Client"
+        "name": "Test Client",
     }
-    
+
     # Act
     user = factory.create_user(user_data, db_session)
-    
+
     # Assert
     assert user.id is not None  # ID should be assigned by the database
     assert isinstance(user.account_id, UUID)
     assert user.user_type == UserType.CLIENT
     assert user.is_staff is False
-    
+
     # Verify user is in the database
     db_user = db_session.get(User, user.id)
     assert db_user is not None
